move data models out of internal package
